refactor(agents): extract aggregate token usage model
