+++ common/src/main/java/com/netflix/conductor/common/run/AggregateTokenUsage.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2026 Conductor Authors.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package com.netflix.conductor.common.run;
+
+/** Token usage summed across a workflow and its sub-workflow tree. */
+public class AggregateTokenUsage {
+
+    private long promptTokens;
+    private long completionTokens;
+    private long totalTokens;
+
+    public AggregateTokenUsage() {}
+
+    public AggregateTokenUsage(long promptTokens, long completionTokens, long totalTokens) {
+        this.promptTokens = promptTokens;
+        this.completionTokens = completionTokens;
+        this.totalTokens = totalTokens;
+    }
+
+    public long getPromptTokens() {
+        return promptTokens;
+    }
+
+    public void setPromptTokens(long promptTokens) {
+        this.promptTokens = promptTokens;
+    }
+
+    public long getCompletionTokens() {
+        return completionTokens;
+    }
+
+    public void setCompletionTokens(long completionTokens) {
+        this.completionTokens = completionTokens;
+    }
+
+    public long getTotalTokens() {
+        return totalTokens;
+    }
+
+    public void setTotalTokens(long totalTokens) {
+        this.totalTokens = totalTokens;
+    }
+}
